refactor(serve,session): split serve.rs monolith into serve/ submodule
serve.rs (2578 lines) was two subsystems in one file. Split into:
- serve/mod.rs — shared types (StreamFrame, AppState, SSE utilities)
- serve/run_session.rs — one-shot POST /messages handlers + RunFn seam
- serve/live_session.rs — bidirectional live-session handlers

Additional structural cleanup:
- LiveSessionControl now owns typed try_set_model / try_get_context_usage /
  send_turn methods; the dispatcher is a clean match on action string
- spawn_frame_forwarder extracted once; both SSE paths share it via closure
- CodexSessionOptions gains with_approval_policy / with_sandbox builders,
  eliminating the dual-default-unwrap pattern
- session.rs execute_new_claude + execute_new_codex collapsed into one code
  path using Arc<Ctrl> + boxed closures

All 24 smoke tests pass; clippy -D warnings clean.
